fix: firestore native sdk
